requirements.txt
fastapi
uvicorn
jinja2
google-generativeai
fpdf
requests
python-multipart
python-dotenv